Fix a softlock and some build warnings.

+++ SonicMania/Objects/FBZ/FoldingPlatform.h
@@ -14,7 +14,7 @@ struct EntityFoldingPlatform {
 
     uint16 interval;
     uint16 intervalOffset;
-    int32 duration;
+    int16 duration;
 };
 
 // Object Struct
